#2270 created test utils file
