+++ src/backend/tests/unmocked/work-package-template.test.ts
@@ -1,34 +1,13 @@
 import prisma from '../../src/prisma/prisma';
 import WorkPackageService from '../../src/services/work-packages.services';
 import { AccessDeniedGuestException, HttpException } from '../../src/utils/errors.utils';
-import { createTestUser } from '../create-test-user';
+import { createTestUser, resetUsers } from '../test-utils';
 import { batman, theVisitor, thomasEmrax } from '../test-data/users.test-data';
 import { WorkPackageTemplate1 } from '../test-data/work-packages.test-data';
 
 describe('Work Package Template Tests', () => {
   beforeEach(async () => {
-    await prisma.material.deleteMany();
-    await prisma.manufacturer.deleteMany();
-    await prisma.material_Type.deleteMany();
-    await prisma.assembly.deleteMany();
-    await prisma.team.deleteMany();
-    await prisma.user_Secure_Settings.deleteMany();
-    await prisma.reimbursement_Product.deleteMany();
-    await prisma.reimbursement_Status.deleteMany();
-    await prisma.reimbursement_Request.deleteMany();
-    await prisma.task.deleteMany();
-    await prisma.stage_Gate_CR.deleteMany();
-    await prisma.activation_CR.deleteMany();
-    await prisma.change.deleteMany();
-    await prisma.proposed_Solution.deleteMany();
-    await prisma.scope_CR_Why.deleteMany();
-    await prisma.scope_CR.deleteMany();
-    await prisma.change_Request.deleteMany();
-    await prisma.link.deleteMany();
-    await prisma.linkType.deleteMany();
-    await prisma.work_Package_Template.deleteMany();
-    await prisma.user_Settings.deleteMany();
-    await prisma.user.deleteMany();
+    await resetUsers();
   });
 
   describe('Get single work package template', () => {
